refactor: update import statement for merge function in CLI module
